refactor: change directory of cargo orders components
