+++ src/components/CargoOrders/OrderCard/OrderActions.tsx
@@ -1,6 +1,6 @@
 import { Link } from 'react-router';
-import { Eye } from "../../assets/Eye.tsx";
-import { formatTimeDiff } from "../../utils/formatters.ts";
+import { Eye } from "../../../assets/Eye.tsx";
+import { formatTimeDiff } from "../../../utils/formatters.ts";
 interface OrderActionsProps {
     inTransit: boolean;
     orderNumber: string | number;
