+++ src/components/CargoOrders/OrderCard/OrderCard.tsx
@@ -1,4 +1,4 @@
-import type { Order } from '../../types/order';
+import type { Order } from '../../../types/order.ts';
 
 import OrderHeader from "./OrderHeader.tsx";
 import OrderStatusBar from "./OrderStatusBar.tsx";
